Projects/Tutor_Agent/requirements.txt
flask
requests
python-dotenv
langchain
google-generativeai
langchain-community
langchain_google_genai